move to QuasiSpec
